Add -D option to interpret X values as timestamps
Add -D option to interpret X values as Unix timestamps (seconds since
1970-01-01 00:00:00 GMT) and label the X axis with GMT dates or times
depending upon the scale.  Each label shows the increment of time at
the scale that fits, while at the left and right edges an additional
label shows the remaining date and time information not included in
the labels.  The format of the labels will vary to show a larger scale
value on the even boundaries.  Examples:

2013                                                           2014
       10/01       12/01       02/01       04/01       06/01

2014-04-25                                               2014-04-26
     22:30     23:00     23:30    04/26    00:30    01:00

2916-08-10 17:55                                   2016-08-10 17:56
  45.000 50.000 55.000 17:56  05.000 10.000 15.000 20.000 25.000