Add -L option to display timestamps in local time not GMT
Note: Accommodating DST changes is a real pain!  We make the grid
increment uneven to maintain the same offset from midnight before and
after the transition.

For both local or GMT times, also avoid making the first X grid point
be further right than needed by making the initialization of Xstart be
separate for the date and non-date cases.