Clip data lines at edges of graph area
Previously when zooming in on a graph the data lines would be drawn
all the way to the edges of the window, overlaying grid labels.  Now
the lines will be clipped to the edges of the grid area.

A call to XSetClipRectangles() to do this was present but ifdef'd out,
perhaps because it caused the sample lines in the legend to disappear
after any redraw.  That problem is fixed by enabling the clipping only
during the drawing of the dataset.

This fixes a long-standing but elusive bug that sometimes lines would
exit through a window edge at some odd angle that did not reflect the
actual position of the next point.  That occurred because the code in
DrawSet() determines which points are visible using the grid boundary
as its visibility limit whereas the lines were still visible in the
areas around the grid.  The line exiting the window would be drawn
toward a different data point as the endpoints of lines that were
completely invisible would be skipped.