Clear window before redrawing for resize
A problem seen on Mac OS and perhaps other environments is that when
resizing the graph window the image from the previous size and the new
size would be superimposed, requiring some other operation to cause
the graph to be redrawn.  Now in ResizeWindow() we clear the window
before redrawing.