When toggling to unhide a dataset, draw it on top
Rather than alway clearing and redrawing the graph when toggling a
dataset to hide or unhide, do the full redraw only when hiding.  When
unhiding a dataset, just draw it on the graph as it exists, which will
put that dataset on top of the others.  This is useful to pull
obscured data forward to make it visible.  Ctrl-L will redraw the
whole graph with the datasets in the original order.

Change Ctrl-L to always clear the window and redraw rather than
redrawing from the cached pixmap because the pixmap may not have a
clean copy (as used to be the case when resizing the window before the
resize was changed to always clear and redraw the window).