Set number precision when X axis is date mode
When -D or -L has been selected to display the X axis as dates, we
need to set XPrecisionOffset in the same manner as would be done if
the date option were not selected, in order to display the large
timestamp values appropriately.  In particular this precision affects
how the X coordinate is shown when holding he left mouse button.  We
reduce by 2 the precision calculated for that operation.