Make toggling of graph outline work
Change the keystroke function 'b' to 'o' to match the command-line
switch -o and rename the corresponding flags.bb to flags.outline
because what this function does is to toggle drawing a border
outline around the graph.  It is not related to the bounding box of
the data.

Also fix the keystroke function implementation so it toggles the flag
correctly and actually works as opposed to doing nothing.