Add a Help window
When the user types ?, Ctrl-H or F1 in the graph window, create a
new window containing text explaining the available mouse functions
and keystroke commands.  Sizing of the window and the tab stop for the
keystroke list are calculated dynamically from the text strings.  The
help window positioned to the right of the graph window.

Add a note advertising this new feature at the bottom of the usage()
output and remove a note there about fonts needing to be fixed width
since the default font is not.