Fix a clipping bug inside the graph area
After 70b303c fixed the longstanding bug of lines drawn at incorrect
angles outside of the graph area, clipping at the boundary of the
graph was enabled.  Then after the change from a percentage padding to
fixed-size padding in ae9cc57b the graph lines in a zoomed image might
be clipped before they reached the edge of the graph.  The problem was
that wi->loX, wi->hiX, wi->loY, wi->hiY are no longer at the outline
boundary of the graph, and instead we need to use wi->UsrOrgX, etc.