Implement pan and zoom with scroll wheel or arrow keys
graf is now able to pan up, down, left and right in a zoomed view
using a mouse scroll wheel or trackpad gesture or using the arrow
keys.  It can also zoom in and out at the cursor using by holding the
shift key while scrolling up and down or pressing the up and down
arrow keys.  This is in addition to the existing zooming feature using
the right mouse button to select a subset area to zoom.

When zooming out, the zoom will stop once the overall bounding box is
reached.  When zooming by dragging a bounding box with the right mouse
button, the bounding box is now constrained to be within the overall
bounding box of the data space.  If X or Y limits are set by command
line options that are smaller than the size of the data space, the
overall bounding box is expanded such that it is possible to zoom out
and see all the data while keeping the aspect ratio imposed by the
limits.

A new X resource "NaturalScroll" was added to allow changing the
panning behavior such that the content of the graph is moved rather
than the viewport of the graph (following the Mac OS nomenclature for
which scrolling direction is "natural").

Also added panning and zooming to the help text and updated README:

- Remove a comment about no panning and add instructions for
  scroll-wheel pan and zoom.

- A few tweaks in the intro attribution paragraph.

- List the available Xresources.