Improve positioning of cursor labels
For the Identify, Distance and Regression operations, if the label
under the cursor would overlap the X axis labels, move the label above
the cursor instead.  Also, when the label is shifted horizontally
because it would extend past the edge of the window, add padding so
that its not drawn right against the window edge.