Enable differentiating multiple datapoints at same location
When looking for the datapoint closest to the cursor for Identify,
Distance, etc., if there are multiple datapoints at the same location
then prefer the lower-numbered one if the cursor is to the left of the
datapoints, or the higher-numbered one if the cursor is to the right.
This allows seeing that there are multiple datapoints there and, if
they are consecutive, how many.