Fix bug when first date grid line rolls to next year
When the X axis is in date mode at a resoution of months, the Xstart
grid line is aligned to start on the next first-of-the-month.  When
that rolls over from December to January, we were incrementing the
year but not resetting the month to January.  This was causing Xstart
to be past the end of the graph, so no X axis was shown at all.