Implement week alignment for date-mode X axis
If the initial grid increment for the date-mode X axis is larger than
5 days but not larger than 14 days, make the increment be 7 days or 14
days and align the Xstart grid line to the next Sunday.  Take
advantage of mktime's flexibility to normalize tm_mday larger than 31
for this.  Similarly for months mode I removed my explicit
normalization code for the year rollover.